requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113 
torchvision==0.11.1+cu113
PyMySQL==1.0.2
scipy
diskcache==5.2.1
uvicorn==0.13.4
numpy==1.21.3
pydantic==1.8.2
pymilvus==2.0.2
# pymilvusdm==2.0
towhee==0.6.0
fastapi==0.65.2
python-multipart==0.0.5
pillow==8.4.0
aiofiles==0.7.0
easydict
opencv-python